requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
alembic==1.11.1
Flask-JWT-Extended==4.5.2
sqlparse==0.4.4
boto3==1.28.2